Remove lib/utils/access
